Q1-601-v2: Defenage - Incorrectly calculating discount for the rule type "Fixed amount for whole cart".
